% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-data.R
\name{pattern_data}
\alias{pattern_data}
\title{Design Pattern for an R Data Package}
\usage{
pattern_data(path,
  title = "What the Package Does (One Line, Title Case)",
  description = "What the package does (one paragraph)",
  license = "What license it uses")
}
\arguments{
\item{path}{Where should the package be located and named?}

\item{title}{What the package does?}

\item{description}{What does the package do?}

\item{license}{What is the package licensed under?}
}
\description{
Setups an R Package for Data Use
}
