% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-pkg-backend.R
\name{pattern_pkg_init}
\alias{pattern_pkg_init}
\title{Package Initialization and Finalization Routines}
\usage{
pattern_pkg_init(path,
  title = "What the Package Does (One Line, Title Case)",
  description = "What the package does (one paragraph)",
  license = "What license it uses")
}
\arguments{
\item{path}{Where should the package be located and named?}

\item{title}{What the package does?}

\item{description}{What does the package do?}

\item{license}{What is the package licensed under?}
}
\description{
Common package initialization features.
}
\keyword{internal}
