% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdesign-package.R
\docType{package}
\name{pkgdesign-package}
\alias{pkgdesign}
\alias{pkgdesign-package}
\title{pkgdesign: R Package Design Patterns}
\description{
Provides wrapper functions that setup an R package based on 
    an acceptable design pattern for the task at hand.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-pkg-examples/pkgdesign}
  \item Report bugs at \url{https://github.com/r-pkg-examples/pkgdesign/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (0000-0003-2826-8458)

}
\keyword{internal}
